MLE_ile_akıllı_sehir/requirements.txt
numpy
scipy.optimize
matplotlib
scipy.stats